Add getEvents to Instances
Instances can now return events. They're fetched on first request, but
cached for subsequent calls. The `events` field has been left public
so that synchronous code can use `events` if they've been fetched.

Currently a call to `refreshFields` unsets `events`. I'm sure that
having the old events is not desirable, but I'm unsure if this is the
correct choice, or if I should be refetching.